Fixed crash on merging of DDE patient without npid.
Retrieving of NPIDs from patient objects returned by DDE is somewhat heuristical
as there is no properly defined structure for a patient in DDE (especially on the identifiers part).
In cases where the NPID is missing the exact structure is highly uncertain.
In some instances the BHT-EMR-API was crashing due to this.